requirements.txt
faiss-cpu
groq
langchain-groq
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
openpyxl
google-cloud-aiplatform>=1.38
gspread
google-auth